% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ConumeeAnno450k}
\alias{ConumeeAnno450k}
\title{ConumeeAnno450k.}
\format{An object of class \code{CNV annotation object} as generated by \code{conumee}.
Refer to \url{http://bioconductor.org/packages/release/bioc/vignettes/conumee/inst/doc/conumee.html}
for more information.}
\usage{
ConumeeAnno450k
}
\description{
Precomputed annotation for the 450K platform used in CNV calling by \code{conumee}.
}
\keyword{datasets}
