% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ConumeeAnnoEpic}
\alias{ConumeeAnnoEpic}
\title{ConumeeAnnoEpic.}
\format{An object of class \code{CNV annotation object} as generated by \code{conumee}.
Refer to \url{http://bioconductor.org/packages/release/bioc/vignettes/conumee/inst/doc/conumee.html}
for more information.}
\usage{
ConumeeAnnoEpic
}
\description{
Precomputed annotation for the EPIC platform used in CNV calling by \code{conumee}.
}
\keyword{datasets}
