% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input.R
\name{get_all}
\alias{get_all}
\alias{get_invalid}
\alias{get_green_only}
\alias{get_red_only}
\alias{get_cases}
\title{Accessor functions for scan_directory() output.}
\usage{
get_all(obj)

get_invalid(obj)

get_green_only(obj)

get_red_only(obj)

get_cases(obj)
}
\arguments{
\item{obj}{Output of \code{scan_directory}}
}
\value{
Vector with file names or basenames (see above)
}
\description{
After scanning a directory with \code{\link{scan_directory()}}, the accessor
functions listed here can be used to extract useful information.
}
\details{
\describe{
  \item{get_all}{Lists all files}
  \item{get_invalid}{Lists all non-idat files}
  \item{get_green_only}{Lists \emph{basenames} with missing RED channel data}
  \item{get_red_only}{Lists \emph{basenames} with missing GREEN channel data}
  \item{get_cases}{Lists \emph{basenames} of complete cases}
}
}
