% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{get_basename}
\alias{get_basename}
\title{Extract basename from IDAT filename.}
\usage{
get_basename(file)
}
\arguments{
\item{file}{Vector of file name(s)}
}
\value{
Vector with basenames
}
\description{
Given the name of an IDAT file (such as \emph{200134080018_R03C01_Grn.idat})
\code{get_basename()} will strip the suffix and return the corresponding basename
(\emph{200134080018_R03C01}).
}
\keyword{internal}
