% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_reports.R
\name{get_filename_extension}
\alias{get_filename_extension}
\title{Extract file name extension}
\usage{
get_filename_extension(filename)
}
\arguments{
\item{filename}{Single filename of class character.}
}
\value{
File name extension.
}
\description{
Extracts file name extension ("." followed by alphanumerical characters at the end of a string)
}
