% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{guess_array_type}
\alias{guess_array_type}
\title{Guess type of DNA methylation array.}
\usage{
guess_array_type(n_probes)
}
\arguments{
\item{n_probes}{Number of probes detected}
}
\value{
Array type, one of:
\itemize{
  \item IlluminaHumanMethylationEPIC
  \item IlluminaHumanMethylation450k
  \item IlluminaHumanMethylation27k
  \item HorvathMammalMethylChip40
  \item unknown
}
}
\description{
Will report the platform of a DNA methylation array given the number
of probes measured. Adapted from an internal function in \code{minfi} package.
Returns \emph{unknown} if array platform cannot be determined.
}
\keyword{internal}
