% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{impute_random}
\alias{impute_random}
\title{Impute missing values.}
\usage{
impute_random(mat)
}
\arguments{
\item{mat}{Matrix (with or without missing data)}
}
\value{
List with two slots:
 \item{imputed}{Updated version of input matrix with NAs replaced}
 \item{n}{Number of data points which were replaced}
}
\description{
This function imputes missing data with data points samples randomly from the non-missing entries.
Returns the input data unchanged if all data are missing.
}
