% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NormalControls450k}
\alias{NormalControls450k}
\title{NormalControls450k.}
\format{An object of class \code{MethylSet} as generated with the \code{minfi}
  package. Preprocessed with function \code{minfi::preprocessIllumina()}.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE86833}
}
\usage{
NormalControls450k
}
\description{
Normal controls (n = 5) used for copynumber calling by the \code{conumee} package.
Data were obtained from NCBI's GEO database, and preprocessed using \code{minfi}.
}
\details{
Note: Normal controls for EPIC data are derived from the same samples.
}
\keyword{datasets}
