% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RFpurify_ABSOLUTE}
\alias{RFpurify_ABSOLUTE}
\title{RFpurify_ABSOLUTE.}
\format{An object of class \code{randomForest}.}
\source{
\url{https://github.com/mwsill/RFpurify}
}
\usage{
RFpurify_ABSOLUTE
}
\description{
Classifier for methylation-based tumor purity prediction trained on ESTIMATE data.
}
\keyword{datasets}
