% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input.R
\name{scan_directory}
\alias{scan_directory}
\title{Scan a directory for IDAT files.}
\usage{
scan_directory(dir)
}
\arguments{
\item{dir}{Directory to scan.}
}
\value{
Returns a list with the following slots:
 \item{all}{List of all files in dir}
 \item{invalid}{List of non-IDAT files}
 \item{green_only}{Basenames of cases with missing RED channel data}
 \item{red_only}{Basenames of cases with missing GREEN channel data}
 \item{cases}{Basenames of cases with correctly paired IDAT files}
}
\description{
\code{scan_directory} scans a directory, detects IDAT and non-IDAT files and
reports basenames for complete cases .
}
\details{
This function scans a directory. Its main purpose is to identify cases
with paired (_Grn.idat and _Red.idat) IDAT files. It will also report
any invalid files (types other than .idat) or cases where one of the
input files is missing.

\strong{Note}: Will report basenames, not filenames for IDAT files

Use accessor functions \code{\link{get_all}}, \code{\link{get_invalid}},
\code{\link{get_cases}}, \code{\link{get_green_only}} and
\code{\link{get_red_only}} to extract data.

This implementation might currently be overkill, but was created with
forward compatibility in mind (might be replaced by a specialized class in
future versions).
}
