% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_reports.R
\name{render_report}
\alias{render_report}
\title{Render diagnostic report for ClassificationCase object}
\usage{
render_report(case, template, out_dir, out_file = NULL, out_type = "pdf")
}
\arguments{
\item{case}{Object of type ClassificationCase.}

\item{template}{Template (.Rmd file) to generate report.}

\item{out_dir}{Output directory of report.}

\item{out_file}{Name of output file, objects basename if unspecified (Default: NULL).}

\item{out_type}{Output file type (html or pdf, Default: pdf)}
}
\value{
Full path of report file.
}
\description{
\strong{Note}: Will report basenames, not filenames for IDAT files
}
\details{
Use this function to render a diagnostic report as PDF after running the full
workflow for a \emph{ClassificationCase} object. Internally, render_report generates
a new environment and and passes the parameters to rmarkdown::render(). Output reports can be
rendered as HTML or PDF files. Please note: if out_file extension does not match out_type,
out_type takes precedence and will change the extension.
}
